bætti við myndum i connection

+++ GUI-verklegtVika3/deleteconnection.cpp
@@ -1,5 +1,6 @@
 #include "deleteconnection.h"
 #include "ui_deleteconnection.h"
+#include <QPixmap>
 
 DeleteConnection::DeleteConnection(QWidget *parent) :
     QDialog(parent),
@@ -7,6 +8,8 @@ DeleteConnection::DeleteConnection(QWidget *parent) :
 {
     setWindowFlags(windowFlags() & ~Qt::WindowContextHelpButtonHint);
     ui->setupUi(this);
+    QPixmap pix("/Users/irisdogg/Desktop/Verklegt-1/GUI-verklegtVika3/deleteConnection.png");
+    ui->deletePic->setPixmap(pix.scaled(300,100));
 }
 
 DeleteConnection::~DeleteConnection()
@@ -20,11 +23,21 @@ void DeleteConnection::on_deleteButton_clicked()
     QString sciID = ui->insertSciID->text();
     QString compID = ui->insertCompID->text();
 
-    int tempSci = sciID.toInt();
-    int tempComp = compID.toInt();
+    if (sciID.isEmpty())
+    {
+        ui->errorCheck->setText("<span style = 'color:red'>Connection failed!</span>");
+    }
 
-    qDebug() << tempSci;
-    qDebug() << tempComp;
+    if (compID.isEmpty())
+    {
+        ui->errorCheck->setText("<span style = 'color:red'>Connection failed!</span>");
+    }
+    else
+    {
+        int tempSci = sciID.toInt();
+        int tempComp = compID.toInt();
 
-    connectionService.deleteLink(tempComp,tempSci);
+        connectionService.deleteLink(tempComp,tempSci);
+        close();
+    }
 }
